requirements.txt
rich
discord.py
python-dotenv
pandas
openpyxl
pytest
pytest-asyncio
psycopg[binary]
psycopg-pool
